refactor sftp
