+++ unstructured/ingest/cli/cmds/fsspec/sftp.py
@@ -7,13 +7,13 @@ from unstructured.ingest.cli.base.src import BaseSrcCmd
 from unstructured.ingest.cli.interfaces import (
     CliConfig,
 )
-from unstructured.ingest.connector.sftp import SftpAccessConfig, SimpleSftpConfig
+from unstructured.ingest.connector.fsspec.sftp import SimpleSftpConfig
 
 CMD_NAME = "sftp"
 
 
 @dataclass
-class SftpCliConfig(SftpAccessConfig, CliConfig):
+class SftpCliConfig(SimpleSftpConfig, CliConfig):
     @staticmethod
     def get_cli_options() -> t.List[click.Option]:
         options = [
@@ -37,7 +37,6 @@ def get_base_src_cmd() -> BaseSrcCmd:
     cmd_cls = BaseSrcCmd(
         cmd_name=CMD_NAME,
         cli_config=SftpCliConfig,
-        addition_configs={"fsspec_config": SimpleSftpConfig},
         is_fsspec=True,
     )
     return cmd_cls
